the_wall/docker/thewall/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
Flask==2.1.1
Flask-JWT-Extended==4.3.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pymongo==4.1.0
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
